# MoveYourBum/MoveYourBum/OpenAPIs/swagger2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MoveYourBumAPI",
    "version": "1.0"
  },
  "paths": {
    "/api/Day": {
      "get": {
        "tags": [
          "Day"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayForView"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayForView"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayForView"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Day"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DayForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DayForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DayForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DayForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayForView"
                }
              }
            }
          }
        }
      }
    },
    "/api/Day/{id}": {
      "get": {
        "tags": [
          "Day"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DayForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayForView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Day"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DayForView"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DayForView"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DayForView"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DayForView"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayForView"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayForView"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Day"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DayForView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "daySchedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayScheduleForView"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DayScheduleForView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idDay": {
            "type": "integer",
            "format": "int32"
          },
          "dayDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "idSchedule": {
            "type": "integer",
            "format": "int32"
          },
          "scheduleName": {
            "type": "string",
            "nullable": true
          },
          "scheduleDescription": {
            "type": "string",
            "nullable": true
          },
          "scheduleExercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleExerciseForView"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ScheduleExerciseForView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "annotation": {
            "type": "string",
            "nullable": true
          },
          "idSchedule": {
            "type": "integer",
            "format": "int32"
          },
          "idExercise": {
            "type": "integer",
            "format": "int32"
          },
          "exerciseName": {
            "type": "string",
            "nullable": true
          },
          "sets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleExerciseSetForView"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ScheduleExerciseSetForView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "plannedReps": {
            "type": "string",
            "nullable": true
          },
          "actualReps": {
            "type": "string",
            "nullable": true
          },
          "weightUsed": {
            "type": "string",
            "nullable": true
          },
          "idScheduleExercise": {
            "type": "integer",
            "format": "int32"
          },
          "scheduleExerciseName": {
            "type": "string",
            "nullable": true
          },
          "idDaySchedule": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "dayScheduleName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}